+++ src/Monify.Console.Tests/Classes/Simple/SimpleForArrayTests/WhenToStringIsCalled.cs
@@ -2,7 +2,7 @@ namespace Monify.Console.Classes.Simple.SimpleForArrayTests;
 
 public static class WhenToStringIsCalled
 {
-    private const string Expected = "SimpleForArray { System.Int32[] }";
+    private const string Expected = "System.Int32[]";
     private static readonly int[] _sampleValue = [1, 2, 3];
 
     [Fact]