+++ src/Monify.Tests/Snippets/Declarations/Classes/Nested.MultiLevel.Expected.NonNullable.cs
@@ -392,7 +392,12 @@ internal static partial class Nested
                                 {
                                     public override string ToString()
                                     {
-                                        return string.Format("InlineStyle {{ {0} }}", _value);
+                                        if (ReferenceEquals(_value, null))
+                                        {
+                                            return string.Empty;
+                                        }
+
+                                        return _value.ToString();
                                     }
                                 }
                             }