+++ src/Monify.Tests/Snippets/Declarations/Classes/Simple.Expected.Nullable.cs
@@ -391,7 +391,12 @@ internal static partial class Simple
                     {
                         public override string ToString()
                         {
-                            return string.Format("Simple {{ {0} }}", _value);
+                            if (ReferenceEquals(_value, null))
+                            {
+                                return string.Empty;
+                            }
+
+                            return _value.ToString();
                         }
                     }