+++ src/Monify.Tests/Snippets/Declarations/Structs/Simple.Expected.NonNullable.cs
@@ -304,7 +304,12 @@ internal static partial class Simple
                     {
                         public override string ToString()
                         {
-                            return string.Format("Simple {{ {0} }}", _value);
+                            if (ReferenceEquals(_value, null))
+                            {
+                                return string.Empty;
+                            }
+
+                            return _value.ToString();
                         }
                     }
                 }